Remaining diff..  Support templateholder creation based on page template!
The `PlaceholderInline` retrieves default values from the parent
(with `get_placeholder_data()`), and passes it to the formset.

The formset `save_new()` method links the Placeholder that was just
created by the `PlaceholderInline`. The placeholder_slot is no longer
needed in the model, it is only needed in the form when adding pages.

Lots of stuff moved to Mixins, to allow reuse among different admin classes. e.g.:
- Both `PlaceholderAdmin` and `PlaceholderFieldAdmin` can create inlines dynamically using `DynamicInlinesAdminMixin`.
- Base class for placeholder admin is `PlaceholderFieldAdminMixin`, allowing both normal `ModelAdmin` and `MPTTModelAdmin` classes to inherit it.
- The ``ContentItemInline`` was moved to a Mixin, so a normal `InlineModelAdmin` and `GenericInlineModelAdmin` can inherit it.

Templates updates
- deal with the dummy/empty item that inline_admin_formset creates.
- deal with initial values in the form, render those instead of `inline_admin_model.original`.