markup plugin: support all languages, via proxy model.
This can also be seen as an experiment, internally the polymorphic model
is still the MarkupItem, so databasewise nothing changed.

Only the method for choosing the language moved to the "add plugin"
select box. When a language is no longer supported, the plugin can't be
displayed anymore (which is fair enough given the dependencies that are
probably removed).