Remove TranslatableAdmin from PlaceholderEditorAdmin, as that affects all projects
Instead, handle the new `post_translation_delete` signal to make sure
that objects associated with a translation will be deleted. This removes
the need to overwrite `delete_model_translation()` in the every admin.

Using multilingual support in the admin is now fully optional;
when a translation system is used in the admin, the contents will adapt
itself to it.